[Update]注文情報のバリデーション設定

+++ app/controllers/public/orders_controller.rb
@@ -23,20 +23,24 @@ class Public::OrdersController < ApplicationController
   end
 
   def create
-    order = Order.new(order_params)
-    order.customer_id = current_customer.id
-    order.save
-    cart_items = current_customer.cart_items
-    cart_items.each do |cart_item|
-      order_item = OrderItem.new(order_item_params)
-      order_item.order_id = order.id
-      order_item.item_id =  cart_item.item_id
-      order_item.price =  cart_item.item.price
-      order_item.amount =  cart_item.amount
-      order_item.save
+    @order = Order.new(order_params)
+    @order.customer_id = current_customer.id
+    if @order.save
+      cart_items = current_customer.cart_items
+      cart_items.each do |cart_item|
+        order_item = OrderItem.new(order_item_params)
+        order_item.order_id = @order.id
+        order_item.item_id =  cart_item.item_id
+        order_item.price =  cart_item.item.price
+        order_item.amount =  cart_item.amount
+        order_item.save
+      end
+      cart_items.destroy_all
+      redirect_to orders_thanks_path
+    else
+      @customer = current_customer
+      render :new
     end
-    cart_items.destroy_all
-    redirect_to orders_thanks_path
   end
 
   def thanks
